Add AppSignal integration
Using AppSignal's `send_error` functionality
( https://docs.appsignal.com/ruby/instrumentation/exception-handling.html#appsignal-send_error )
we can send an error with tags and namespace variables to AppSignal.

Note: The namespace is optional, but should contain a value if passed
(see :https://github.com/appsignal/appsignal-ruby/blob/main/lib/appsignal/helpers/instrumentation.rb#L197)
this means that we shouldn't set it to `nil`
if a namespace is not given in either the options or arguments.
To do this, we splat the array after compacting it, effectively removing the 3rd argument if it's `nil`.